fix: comfigSetDefault action dispatched from resetSettings in SaveResetApplyPreset in ReaderSettings.tsx
fix: create a copy of the color obj and not copy by reference for the
keys : ttsHighlightColor, ttsHighlightColor_WORD, annotation_defaultColor

fix: configCopy variable assignation , previously was not deeply copied

tested on ttsHighlightColor and settings/menu section on epub navigator
tested on pdf/divina -> reset to defaultSate do not persist
settings/menu section
